@@ -305,13 +305,13 @@ class _MyHomePageState extends State<MyHomePage> {
               FlatButton(
                   child: const Text('PICK'),
                   onPressed: () {
-                    double width = maxWidthController.text.length > 0
+                    double width = maxWidthController.text.isNotEmpty
                         ? double.parse(maxWidthController.text)
                         : null;
-                    double height = maxHeightController.text.length > 0
+                    double height = maxHeightController.text.isNotEmpty
                         ? double.parse(maxHeightController.text)
                         : null;
-                    int quality = qualityController.text.length > 0
+                    int quality = qualityController.text.isNotEmpty
                         ? int.parse(qualityController.text)
                         : null;
                     onPick(width, height, quality);
