@@ -4,9 +4,11 @@
 
 package io.flutter.plugins.inapppurchase;
 
+import static io.flutter.plugins.inapppurchase.Translator.fromBillingResult;
 import static io.flutter.plugins.inapppurchase.Translator.fromPurchasesList;
 
 import androidx.annotation.Nullable;
+import com.android.billingclient.api.BillingResult;
 import com.android.billingclient.api.Purchase;
 import com.android.billingclient.api.PurchasesUpdatedListener;
 import io.flutter.plugin.common.MethodChannel;
@@ -22,9 +24,10 @@ class PluginPurchaseListener implements PurchasesUpdatedListener {
   }
 
   @Override
-  public void onPurchasesUpdated(int responseCode, @Nullable List<Purchase> purchases) {
+  public void onPurchasesUpdated(BillingResult billingResult, @Nullable List<Purchase> purchases) {
     final Map<String, Object> callbackArgs = new HashMap<>();
-    callbackArgs.put("responseCode", responseCode);
+    callbackArgs.put("billingResult", fromBillingResult(billingResult));
+    callbackArgs.put("responseCode", billingResult.getResponseCode());
     callbackArgs.put("purchasesList", fromPurchasesList(purchases));
     channel.invokeMethod(InAppPurchasePlugin.MethodNames.ON_PURCHASES_UPDATED, callbackArgs);
   }
