@@ -16,6 +16,10 @@ PurchaseWrapper _$PurchaseWrapperFromJson(Map json) {
     sku: json['sku'] as String,
     isAutoRenewing: json['isAutoRenewing'] as bool,
     originalJson: json['originalJson'] as String,
+    developerPayload: json['developerPayload'] as String,
+    isAcknowledged: json['isAcknowledged'] as bool,
+    purchaseState:
+        const PurchaseStateConverter().fromJson(json['purchaseState'] as int),
   );
 }
 
@@ -29,11 +33,39 @@ Map<String, dynamic> _$PurchaseWrapperToJson(PurchaseWrapper instance) =>
       'sku': instance.sku,
       'isAutoRenewing': instance.isAutoRenewing,
       'originalJson': instance.originalJson,
+      'developerPayload': instance.developerPayload,
+      'isAcknowledged': instance.isAcknowledged,
+      'purchaseState':
+          const PurchaseStateConverter().toJson(instance.purchaseState),
+    };
+
+PurchaseHistoryRecordWrapper _$PurchaseHistoryRecordWrapperFromJson(Map json) {
+  return PurchaseHistoryRecordWrapper(
+    purchaseTime: json['purchaseTime'] as int,
+    purchaseToken: json['purchaseToken'] as String,
+    signature: json['signature'] as String,
+    sku: json['sku'] as String,
+    originalJson: json['originalJson'] as String,
+    developerPayload: json['developerPayload'] as String,
+  );
+}
+
+Map<String, dynamic> _$PurchaseHistoryRecordWrapperToJson(
+        PurchaseHistoryRecordWrapper instance) =>
+    <String, dynamic>{
+      'purchaseTime': instance.purchaseTime,
+      'purchaseToken': instance.purchaseToken,
+      'signature': instance.signature,
+      'sku': instance.sku,
+      'originalJson': instance.originalJson,
+      'developerPayload': instance.developerPayload,
     };
 
 PurchasesResultWrapper _$PurchasesResultWrapperFromJson(Map json) {
   return PurchasesResultWrapper(
-    responseCode: _$enumDecode(_$BillingResponseEnumMap, json['responseCode']),
+    responseCode:
+        const BillingResponseConverter().fromJson(json['responseCode'] as int),
+    billingResult: BillingResultWrapper.fromJson(json['billingResult'] as Map),
     purchasesList: (json['purchasesList'] as List)
         .map((e) => PurchaseWrapper.fromJson(e as Map))
         .toList(),
@@ -43,41 +75,24 @@ PurchasesResultWrapper _$PurchasesResultWrapperFromJson(Map json) {
 Map<String, dynamic> _$PurchasesResultWrapperToJson(
         PurchasesResultWrapper instance) =>
     <String, dynamic>{
-      'responseCode': _$BillingResponseEnumMap[instance.responseCode],
+      'billingResult': instance.billingResult,
+      'responseCode':
+          const BillingResponseConverter().toJson(instance.responseCode),
       'purchasesList': instance.purchasesList,
     };
 
-T _$enumDecode<T>(
-  Map<T, dynamic> enumValues,
-  dynamic source, {
-  T unknownValue,
-}) {
-  if (source == null) {
-    throw ArgumentError('A value must be provided. Supported values: '
-        '${enumValues.values.join(', ')}');
-  }
-
-  final value = enumValues.entries
-      .singleWhere((e) => e.value == source, orElse: () => null)
-      ?.key;
-
-  if (value == null && unknownValue == null) {
-    throw ArgumentError('`$source` is not one of the supported values: '
-        '${enumValues.values.join(', ')}');
-  }
-  return value ?? unknownValue;
+PurchasesHistoryResult _$PurchasesHistoryResultFromJson(Map json) {
+  return PurchasesHistoryResult(
+    billingResult: BillingResultWrapper.fromJson(json['billingResult'] as Map),
+    purchaseHistoryRecordList: (json['purchaseHistoryRecordList'] as List)
+        .map((e) => PurchaseHistoryRecordWrapper.fromJson(e as Map))
+        .toList(),
+  );
 }
 
-const _$BillingResponseEnumMap = {
-  BillingResponse.featureNotSupported: -2,
-  BillingResponse.serviceDisconnected: -1,
-  BillingResponse.ok: 0,
-  BillingResponse.userCanceled: 1,
-  BillingResponse.serviceUnavailable: 2,
-  BillingResponse.billingUnavailable: 3,
-  BillingResponse.itemUnavailable: 4,
-  BillingResponse.developerError: 5,
-  BillingResponse.error: 6,
-  BillingResponse.itemAlreadyOwned: 7,
-  BillingResponse.itemNotOwned: 8,
-};
+Map<String, dynamic> _$PurchasesHistoryResultToJson(
+        PurchasesHistoryResult instance) =>
+    <String, dynamic>{
+      'billingResult': instance.billingResult,
+      'purchaseHistoryRecordList': instance.purchaseHistoryRecordList,
+    };
