@@ -20,8 +20,10 @@ SkuDetailsWrapper _$SkuDetailsWrapperFromJson(Map json) {
     sku: json['sku'] as String,
     subscriptionPeriod: json['subscriptionPeriod'] as String,
     title: json['title'] as String,
-    type: _$enumDecode(_$SkuTypeEnumMap, json['type']),
+    type: const SkuTypeConverter().fromJson(json['type'] as String),
     isRewarded: json['isRewarded'] as bool,
+    originalPrice: json['originalPrice'] as String,
+    originalPriceAmountMicros: json['originalPriceAmountMicros'] as int,
   );
 }
 
@@ -39,39 +41,15 @@ Map<String, dynamic> _$SkuDetailsWrapperToJson(SkuDetailsWrapper instance) =>
       'sku': instance.sku,
       'subscriptionPeriod': instance.subscriptionPeriod,
       'title': instance.title,
-      'type': _$SkuTypeEnumMap[instance.type],
+      'type': const SkuTypeConverter().toJson(instance.type),
       'isRewarded': instance.isRewarded,
+      'originalPrice': instance.originalPrice,
+      'originalPriceAmountMicros': instance.originalPriceAmountMicros,
     };
-
-T _$enumDecode<T>(
-  Map<T, dynamic> enumValues,
-  dynamic source, {
-  T unknownValue,
-}) {
-  if (source == null) {
-    throw ArgumentError('A value must be provided. Supported values: '
-        '${enumValues.values.join(', ')}');
-  }
-
-  final value = enumValues.entries
-      .singleWhere((e) => e.value == source, orElse: () => null)
-      ?.key;
-
-  if (value == null && unknownValue == null) {
-    throw ArgumentError('`$source` is not one of the supported values: '
-        '${enumValues.values.join(', ')}');
-  }
-  return value ?? unknownValue;
-}
-
-const _$SkuTypeEnumMap = {
-  SkuType.inapp: 'inapp',
-  SkuType.subs: 'subs',
-};
 
 SkuDetailsResponseWrapper _$SkuDetailsResponseWrapperFromJson(Map json) {
   return SkuDetailsResponseWrapper(
-    responseCode: _$enumDecode(_$BillingResponseEnumMap, json['responseCode']),
+    billingResult: BillingResultWrapper.fromJson(json['billingResult'] as Map),
     skuDetailsList: (json['skuDetailsList'] as List)
         .map((e) => SkuDetailsWrapper.fromJson(e as Map))
         .toList(),
@@ -81,20 +59,22 @@ SkuDetailsResponseWrapper _$SkuDetailsResponseWrapperFromJson(Map json) {
 Map<String, dynamic> _$SkuDetailsResponseWrapperToJson(
         SkuDetailsResponseWrapper instance) =>
     <String, dynamic>{
-      'responseCode': _$BillingResponseEnumMap[instance.responseCode],
+      'billingResult': instance.billingResult,
       'skuDetailsList': instance.skuDetailsList,
     };
 
-const _$BillingResponseEnumMap = {
-  BillingResponse.featureNotSupported: -2,
-  BillingResponse.serviceDisconnected: -1,
-  BillingResponse.ok: 0,
-  BillingResponse.userCanceled: 1,
-  BillingResponse.serviceUnavailable: 2,
-  BillingResponse.billingUnavailable: 3,
-  BillingResponse.itemUnavailable: 4,
-  BillingResponse.developerError: 5,
-  BillingResponse.error: 6,
-  BillingResponse.itemAlreadyOwned: 7,
-  BillingResponse.itemNotOwned: 8,
-};
+BillingResultWrapper _$BillingResultWrapperFromJson(Map json) {
+  return BillingResultWrapper(
+    responseCode:
+        const BillingResponseConverter().fromJson(json['responseCode'] as int),
+    debugMessage: json['debugMessage'] as String,
+  );
+}
+
+Map<String, dynamic> _$BillingResultWrapperToJson(
+        BillingResultWrapper instance) =>
+    <String, dynamic>{
+      'responseCode':
+          const BillingResponseConverter().toJson(instance.responseCode),
+      'debugMessage': instance.debugMessage,
+    };
