@@ -35,10 +35,15 @@ void main() {
 
   setUp(() {
     WidgetsFlutterBinding.ensureInitialized();
+    const String debugMessage = 'dummy message';
+    final BillingResponse responseCode = BillingResponse.ok;
+    final BillingResultWrapper expectedBillingResult = BillingResultWrapper(
+        responseCode: responseCode, debugMessage: debugMessage);
     stubPlatform.addResponse(
         name: startConnectionCall,
-        value: BillingResponseConverter().toJson(BillingResponse.ok));
+        value: buildBillingResultMap(expectedBillingResult));
     stubPlatform.addResponse(name: endConnectionCall, value: null);
+    InAppPurchaseConnection.enablePendingPurchases();
     connection = GooglePlayConnection.instance;
   });
 
@@ -82,10 +87,13 @@ void main() {
         'BillingClient#querySkuDetailsAsync(SkuDetailsParams, SkuDetailsResponseListener)';
 
     test('handles empty skuDetails', () async {
-      final BillingResponse responseCode = BillingResponse.developerError;
-      stubPlatform.addResponse(name: queryMethodName, value: <dynamic, dynamic>{
-        'responseCode': BillingResponseConverter().toJson(responseCode),
-        'skuDetailsList': <Map<String, dynamic>>[]
+      const String debugMessage = 'dummy message';
+      final BillingResponse responseCode = BillingResponse.ok;
+      final BillingResultWrapper expectedBillingResult = BillingResultWrapper(
+          responseCode: responseCode, debugMessage: debugMessage);
+      stubPlatform.addResponse(name: queryMethodName, value: <String, dynamic>{
+        'billingResult': buildBillingResultMap(expectedBillingResult),
+        'skuDetailsList': [],
       });
 
       final ProductDetailsResponse response =
@@ -94,9 +102,12 @@ void main() {
     });
 
     test('should get correct product details', () async {
+      const String debugMessage = 'dummy message';
       final BillingResponse responseCode = BillingResponse.ok;
+      final BillingResultWrapper expectedBillingResult = BillingResultWrapper(
+          responseCode: responseCode, debugMessage: debugMessage);
       stubPlatform.addResponse(name: queryMethodName, value: <String, dynamic>{
-        'responseCode': BillingResponseConverter().toJson(responseCode),
+        'billingResult': buildBillingResultMap(expectedBillingResult),
         'skuDetailsList': <Map<String, dynamic>>[buildSkuMap(dummySkuDetails)]
       });
       // Since queryProductDetails makes 2 platform method calls (one for each SkuType), the result will contain 2 dummyWrapper instead
@@ -110,9 +121,12 @@ void main() {
     });
 
     test('should get the correct notFoundIDs', () async {
+      const String debugMessage = 'dummy message';
       final BillingResponse responseCode = BillingResponse.ok;
+      final BillingResultWrapper expectedBillingResult = BillingResultWrapper(
+          responseCode: responseCode, debugMessage: debugMessage);
       stubPlatform.addResponse(name: queryMethodName, value: <String, dynamic>{
-        'responseCode': BillingResponseConverter().toJson(responseCode),
+        'billingResult': buildBillingResultMap(expectedBillingResult),
         'skuDetailsList': <Map<String, dynamic>>[buildSkuMap(dummySkuDetails)]
       });
       // Since queryProductDetails makes 2 platform method calls (one for each SkuType), the result will contain 2 dummyWrapper instead
@@ -157,8 +171,13 @@ void main() {
   group('queryPurchaseDetails', () {
     const String queryMethodName = 'BillingClient#queryPurchases(String)';
     test('handles error', () async {
+      const String debugMessage = 'dummy message';
       final BillingResponse responseCode = BillingResponse.developerError;
+      final BillingResultWrapper expectedBillingResult = BillingResultWrapper(
+          responseCode: responseCode, debugMessage: debugMessage);
+
       stubPlatform.addResponse(name: queryMethodName, value: <dynamic, dynamic>{
+        'billingResult': buildBillingResultMap(expectedBillingResult),
         'responseCode': BillingResponseConverter().toJson(responseCode),
         'purchasesList': <Map<String, dynamic>>[]
       });
@@ -170,8 +189,13 @@ void main() {
     });
 
     test('returns SkuDetailsResponseWrapper', () async {
+      const String debugMessage = 'dummy message';
       final BillingResponse responseCode = BillingResponse.ok;
+      final BillingResultWrapper expectedBillingResult = BillingResultWrapper(
+          responseCode: responseCode, debugMessage: debugMessage);
+
       stubPlatform.addResponse(name: queryMethodName, value: <String, dynamic>{
+        'billingResult': buildBillingResultMap(expectedBillingResult),
         'responseCode': BillingResponseConverter().toJson(responseCode),
         'purchasesList': <Map<String, dynamic>>[
           buildPurchaseMap(dummyPurchase),
@@ -187,11 +211,16 @@ void main() {
     });
 
     test('should store platform exception in the response', () async {
+      const String debugMessage = 'dummy message';
+
       final BillingResponse responseCode = BillingResponse.developerError;
+      final BillingResultWrapper expectedBillingResult = BillingResultWrapper(
+          responseCode: responseCode, debugMessage: debugMessage);
       stubPlatform.addResponse(
           name: queryMethodName,
           value: <dynamic, dynamic>{
             'responseCode': BillingResponseConverter().toJson(responseCode),
+            'billingResult': buildBillingResultMap(expectedBillingResult),
             'purchasesList': <Map<String, dynamic>>[]
           },
           additionalStepBeforeReturn: (_) {
@@ -225,13 +254,18 @@ void main() {
     test('buy non consumable, serializes and deserializes data', () async {
       final SkuDetailsWrapper skuDetails = dummySkuDetails;
       final String accountId = "hashedAccountId";
+      const String debugMessage = 'dummy message';
       final BillingResponse sentCode = BillingResponse.ok;
+      final BillingResultWrapper expectedBillingResult = BillingResultWrapper(
+          responseCode: sentCode, debugMessage: debugMessage);
+
       stubPlatform.addResponse(
           name: launchMethodName,
-          value: BillingResponseConverter().toJson(sentCode),
+          value: buildBillingResultMap(expectedBillingResult),
           additionalStepBeforeReturn: (_) {
             // Mock java update purchase callback.
             MethodCall call = MethodCall(kOnPurchasesUpdated, {
+              'billingResult': buildBillingResultMap(expectedBillingResult),
               'responseCode': BillingResponseConverter().toJson(sentCode),
               'purchasesList': [
                 {
@@ -242,7 +276,10 @@ void main() {
                   'purchaseTime': 1231231231,
                   'purchaseToken': "token",
                   'signature': 'sign',
-                  'originalJson': 'json'
+                  'originalJson': 'json',
+                  'developerPayload': 'dummy payload',
+                  'isAcknowledged': true,
+                  'purchaseState': 1,
                 }
               ]
             });
@@ -274,13 +311,18 @@ void main() {
     test('handles an error with an empty purchases list', () async {
       final SkuDetailsWrapper skuDetails = dummySkuDetails;
       final String accountId = "hashedAccountId";
+      const String debugMessage = 'dummy message';
       final BillingResponse sentCode = BillingResponse.error;
+      final BillingResultWrapper expectedBillingResult = BillingResultWrapper(
+          responseCode: sentCode, debugMessage: debugMessage);
+
       stubPlatform.addResponse(
           name: launchMethodName,
-          value: BillingResponseConverter().toJson(sentCode),
+          value: buildBillingResultMap(expectedBillingResult),
           additionalStepBeforeReturn: (_) {
             // Mock java update purchase callback.
             MethodCall call = MethodCall(kOnPurchasesUpdated, {
+              'billingResult': buildBillingResultMap(expectedBillingResult),
               'responseCode': BillingResponseConverter().toJson(sentCode),
               'purchasesList': []
             });
@@ -313,13 +355,18 @@ void main() {
         () async {
       final SkuDetailsWrapper skuDetails = dummySkuDetails;
       final String accountId = "hashedAccountId";
+      const String debugMessage = 'dummy message';
       final BillingResponse sentCode = BillingResponse.ok;
+      final BillingResultWrapper expectedBillingResult = BillingResultWrapper(
+          responseCode: sentCode, debugMessage: debugMessage);
+
       stubPlatform.addResponse(
           name: launchMethodName,
-          value: BillingResponseConverter().toJson(sentCode),
+          value: buildBillingResultMap(expectedBillingResult),
           additionalStepBeforeReturn: (_) {
             // Mock java update purchase callback.
             MethodCall call = MethodCall(kOnPurchasesUpdated, {
+              'billingResult': buildBillingResultMap(expectedBillingResult),
               'responseCode': BillingResponseConverter().toJson(sentCode),
               'purchasesList': [
                 {
@@ -330,7 +377,10 @@ void main() {
                   'purchaseTime': 1231231231,
                   'purchaseToken': "token",
                   'signature': 'sign',
-                  'originalJson': 'json'
+                  'originalJson': 'json',
+                  'developerPayload': 'dummy payload',
+                  'isAcknowledged': true,
+                  'purchaseState': 1,
                 }
               ]
             });
@@ -339,9 +389,12 @@ void main() {
       Completer consumeCompleter = Completer();
       // adding call back for consume purchase
       final BillingResponse expectedCode = BillingResponse.ok;
+      final BillingResultWrapper expectedBillingResultForConsume =
+          BillingResultWrapper(
+              responseCode: expectedCode, debugMessage: debugMessage);
       stubPlatform.addResponse(
           name: consumeMethodName,
-          value: BillingResponseConverter().toJson(expectedCode),
+          value: buildBillingResultMap(expectedBillingResultForConsume),
           additionalStepBeforeReturn: (dynamic args) {
             String purchaseToken = args['purchaseToken'];
             consumeCompleter.complete((purchaseToken));
@@ -374,10 +427,13 @@ void main() {
 
     test('buyNonConsumable propagates failures to launch the billing flow',
         () async {
+      const String debugMessage = 'dummy message';
       final BillingResponse sentCode = BillingResponse.error;
+      final BillingResultWrapper expectedBillingResult = BillingResultWrapper(
+          responseCode: sentCode, debugMessage: debugMessage);
       stubPlatform.addResponse(
           name: launchMethodName,
-          value: BillingResponseConverter().toJson(sentCode));
+          value: buildBillingResultMap(expectedBillingResult));
 
       final bool result = await GooglePlayConnection.instance.buyNonConsumable(
           purchaseParam: PurchaseParam(
@@ -389,10 +445,14 @@ void main() {
 
     test('buyConsumable propagates failures to launch the billing flow',
         () async {
-      final BillingResponse sentCode = BillingResponse.error;
+      const String debugMessage = 'dummy message';
+      final BillingResponse sentCode = BillingResponse.developerError;
+      final BillingResultWrapper expectedBillingResult = BillingResultWrapper(
+          responseCode: sentCode, debugMessage: debugMessage);
       stubPlatform.addResponse(
-          name: launchMethodName,
-          value: BillingResponseConverter().toJson(sentCode));
+        name: launchMethodName,
+        value: buildBillingResultMap(expectedBillingResult),
+      );
 
       final bool result = await GooglePlayConnection.instance.buyConsumable(
           purchaseParam: PurchaseParam(
@@ -405,13 +465,17 @@ void main() {
     test('adds consumption failures to PurchaseDetails objects', () async {
       final SkuDetailsWrapper skuDetails = dummySkuDetails;
       final String accountId = "hashedAccountId";
+      const String debugMessage = 'dummy message';
       final BillingResponse sentCode = BillingResponse.ok;
+      final BillingResultWrapper expectedBillingResult = BillingResultWrapper(
+          responseCode: sentCode, debugMessage: debugMessage);
       stubPlatform.addResponse(
           name: launchMethodName,
-          value: BillingResponseConverter().toJson(sentCode),
+          value: buildBillingResultMap(expectedBillingResult),
           additionalStepBeforeReturn: (_) {
             // Mock java update purchase callback.
             MethodCall call = MethodCall(kOnPurchasesUpdated, {
+              'billingResult': buildBillingResultMap(expectedBillingResult),
               'responseCode': BillingResponseConverter().toJson(sentCode),
               'purchasesList': [
                 {
@@ -422,7 +486,10 @@ void main() {
                   'purchaseTime': 1231231231,
                   'purchaseToken': "token",
                   'signature': 'sign',
-                  'originalJson': 'json'
+                  'originalJson': 'json',
+                  'developerPayload': 'dummy payload',
+                  'isAcknowledged': true,
+                  'purchaseState': 1,
                 }
               ]
             });
@@ -431,12 +498,15 @@ void main() {
       Completer consumeCompleter = Completer();
       // adding call back for consume purchase
       final BillingResponse expectedCode = BillingResponse.error;
+      final BillingResultWrapper expectedBillingResultForConsume =
+          BillingResultWrapper(
+              responseCode: expectedCode, debugMessage: debugMessage);
       stubPlatform.addResponse(
           name: consumeMethodName,
-          value: BillingResponseConverter().toJson(expectedCode),
+          value: buildBillingResultMap(expectedBillingResultForConsume),
           additionalStepBeforeReturn: (dynamic args) {
             String purchaseToken = args['purchaseToken'];
-            consumeCompleter.complete((purchaseToken));
+            consumeCompleter.complete(purchaseToken);
           });
 
       Completer completer = Completer();
@@ -469,13 +539,18 @@ void main() {
         () async {
       final SkuDetailsWrapper skuDetails = dummySkuDetails;
       final String accountId = "hashedAccountId";
-      final BillingResponse sentCode = BillingResponse.ok;
+      const String debugMessage = 'dummy message';
+      final BillingResponse sentCode = BillingResponse.developerError;
+      final BillingResultWrapper expectedBillingResult = BillingResultWrapper(
+          responseCode: sentCode, debugMessage: debugMessage);
+
       stubPlatform.addResponse(
           name: launchMethodName,
-          value: BillingResponseConverter().toJson(sentCode),
+          value: buildBillingResultMap(expectedBillingResult),
           additionalStepBeforeReturn: (_) {
             // Mock java update purchase callback.
             MethodCall call = MethodCall(kOnPurchasesUpdated, {
+              'billingResult': buildBillingResultMap(expectedBillingResult),
               'responseCode': BillingResponseConverter().toJson(sentCode),
               'purchasesList': [
                 {
@@ -486,7 +561,10 @@ void main() {
                   'purchaseTime': 1231231231,
                   'purchaseToken': "token",
                   'signature': 'sign',
-                  'originalJson': 'json'
+                  'originalJson': 'json',
+                  'developerPayload': 'dummy payload',
+                  'isAcknowledged': true,
+                  'purchaseState': 1,
                 }
               ]
             });
@@ -495,9 +573,12 @@ void main() {
       Completer consumeCompleter = Completer();
       // adding call back for consume purchase
       final BillingResponse expectedCode = BillingResponse.ok;
+      final BillingResultWrapper expectedBillingResultForConsume =
+          BillingResultWrapper(
+              responseCode: expectedCode, debugMessage: debugMessage);
       stubPlatform.addResponse(
           name: consumeMethodName,
-          value: BillingResponseConverter().toJson(expectedCode),
+          value: buildBillingResultMap(expectedBillingResultForConsume),
           additionalStepBeforeReturn: (dynamic args) {
             String purchaseToken = args['purchaseToken'];
             consumeCompleter.complete((purchaseToken));
@@ -524,20 +605,50 @@ void main() {
         'BillingClient#consumeAsync(String, ConsumeResponseListener)';
     test('consume purchase async success', () async {
       final BillingResponse expectedCode = BillingResponse.ok;
+      const String debugMessage = 'dummy message';
+      final BillingResultWrapper expectedBillingResult = BillingResultWrapper(
+          responseCode: expectedCode, debugMessage: debugMessage);
       stubPlatform.addResponse(
-          name: consumeMethodName,
-          value: BillingResponseConverter().toJson(expectedCode));
+        name: consumeMethodName,
+        value: buildBillingResultMap(expectedBillingResult),
+      );
+      final BillingResultWrapper billingResultWrapper =
+          await GooglePlayConnection.instance
+              .consumePurchase(PurchaseDetails.fromPurchase(dummyPurchase));
 
-      final BillingResponse responseCode = await GooglePlayConnection.instance
-          .consumePurchase(PurchaseDetails.fromPurchase(dummyPurchase));
-
-      expect(responseCode, equals(expectedCode));
+      expect(billingResultWrapper, equals(expectedBillingResult));
     });
   });
 
   group('complete purchase', () {
-    test('calling complete purchase on android should throw', () async {
-      expect(() => connection.completePurchase(null), throwsUnsupportedError);
+    const String completeMethodName =
+        'BillingClient#(AcknowledgePurchaseParams params, (AcknowledgePurchaseParams, AcknowledgePurchaseResponseListener)';
+    test('complete purchase success', () async {
+      final BillingResponse expectedCode = BillingResponse.ok;
+      const String debugMessage = 'dummy message';
+      final BillingResultWrapper expectedBillingResult = BillingResultWrapper(
+          responseCode: expectedCode, debugMessage: debugMessage);
+      stubPlatform.addResponse(
+        name: completeMethodName,
+        value: buildBillingResultMap(expectedBillingResult),
+      );
+      PurchaseDetails purchaseDetails =
+          PurchaseDetails.fromPurchase(dummyUnacknowledgedPurchase);
+      Completer completer = Completer();
+      purchaseDetails.status = PurchaseStatus.purchased;
+      if (purchaseDetails.pendingCompletePurchase) {
+        final BillingResultWrapper billingResultWrapper =
+            await GooglePlayConnection.instance.completePurchase(
+                purchaseDetails,
+                developerPayload: 'dummy payload');
+        print('pending ${billingResultWrapper.responseCode}');
+        print('expectedBillingResult ${expectedBillingResult.responseCode}');
+        print('pending ${billingResultWrapper.debugMessage}');
+        print('expectedBillingResult ${expectedBillingResult.debugMessage}');
+        expect(billingResultWrapper, equals(expectedBillingResult));
+        completer.complete(billingResultWrapper);
+      }
+      expect(await completer.future, equals(expectedBillingResult));
     });
   });
 }
