@@ -5,13 +5,16 @@
 package io.flutter.plugins.webviewflutterexample;
 
 import android.os.Bundle;
+import dev.flutter.plugins.e2e.E2EPlugin;
 import io.flutter.app.FlutterActivity;
-import io.flutter.plugins.GeneratedPluginRegistrant;
+import io.flutter.plugins.webviewflutter.WebViewFlutterPlugin;
 
 public class EmbeddingV1Activity extends FlutterActivity {
   @Override
   protected void onCreate(Bundle savedInstanceState) {
     super.onCreate(savedInstanceState);
-    GeneratedPluginRegistrant.registerWith(this);
+    E2EPlugin.registerWith(registrarFor("dev.flutter.plugins.e2e.E2EPlugin"));
+    WebViewFlutterPlugin.registerWith(
+        registrarFor("io.flutter.plugins.webviewflutter.WebViewFlutterPlugin"));
   }
 }
